Move UI components to a separate module
