+++ game-components/src/main/java/io/github/defective4/javajam/tuisweeper/components/sfx/SFXListSelectDialogBuilder.java
@@ -1,4 +1,4 @@
-package io.github.defective4.javajam.tuisweeper.core.sfx;
+package io.github.defective4.javajam.tuisweeper.components.sfx;
 
 import com.googlecode.lanterna.gui2.*;
 import com.googlecode.lanterna.gui2.dialogs.ListSelectDialog;
