+++ game-components/src/main/java/io/github/defective4/javajam/tuisweeper/components/sfx/SFXRadioBoxList.java
@@ -1,7 +1,7 @@
-package io.github.defective4.javajam.tuisweeper.core.sfx;
+package io.github.defective4.javajam.tuisweeper.components.sfx;
 
-import com.googlecode.lanterna.gui2.ComboBox;
 import com.googlecode.lanterna.gui2.Interactable;
+import com.googlecode.lanterna.gui2.RadioBoxList;
 import com.googlecode.lanterna.input.KeyStroke;
 import com.googlecode.lanterna.input.KeyType;
 
@@ -9,20 +9,11 @@ import com.googlecode.lanterna.input.KeyType;
  * @param <T>
  * @author Defective
  */
-public class SFXComboBox<T> extends ComboBox<T> {
-
+public class SFXRadioBoxList<T> extends RadioBoxList<T> {
     private final SFXEngine sfx;
 
-    @SafeVarargs
-    public SFXComboBox(SFXEngine sfx, T... items) {
-        super(items);
+    public SFXRadioBoxList(SFXEngine sfx) {
         this.sfx = sfx;
-    }
-
-    @Override
-    protected void afterEnterFocus(FocusChangeDirection direction, Interactable previouslyInFocus) {
-        super.afterEnterFocus(direction, previouslyInFocus);
-        sfx.play("focus");
     }
 
     @Override
@@ -32,5 +23,11 @@ public class SFXComboBox<T> extends ComboBox<T> {
         else if (keyStroke.getKeyType() == KeyType.Enter || keyStroke.getKeyType() == KeyType.Character && keyStroke.getCharacter() == ' ')
             sfx.play("confirm");
         return res;
+    }
+
+    @Override
+    protected void afterEnterFocus(FocusChangeDirection direction, Interactable previouslyInFocus) {
+        super.afterEnterFocus(direction, previouslyInFocus);
+        sfx.play("focus");
     }
 }
