+++ game-components/src/main/java/io/github/defective4/javajam/tuisweeper/components/ui/FilteredActionListBox.java
@@ -1,4 +1,4 @@
-package io.github.defective4.javajam.tuisweeper.core.ui;
+package io.github.defective4.javajam.tuisweeper.components.ui;
 
 import com.googlecode.lanterna.TerminalSize;
 import com.googlecode.lanterna.gui2.ActionListBox;
