+++ pom.xml
@@ -17,6 +17,7 @@
     <modules>
         <module>tui-mines-game</module>
         <module>discord-presence-min</module>
+        <module>game-components</module>
     </modules>
 
     <properties>
